apps/settings.py
import datetime
import importlib.metadata
# → Imports Python's standard library module for accessing package metadata.
# What it does: Provides information about installed Python packages.
import json
# → Imports Python's built-in JSON (JavaScript Object Notation) module for working with JSON data.
# What it does: Converts between JSON strings and Python objects.
import logging
# → Imports Python's built-in logging module for recording events, errors, and debug information.
# What it does: Provides a flexible framework for generating log messages from your application.
import os
# → Imports Python's built-in operating system interface module for interacting with the underlying operating system.

# What it does: Provides functions to work with files, directories, environment variables, and system paths.
import os.path
# → Imports the path submodule from Python's os module specifically for path manipulation functions.
# Note: This is often done as import os (which includes os.path), but import os.path explicitly imports just the path utilities.
import warnings
# → Imports Python's built-in warnings module for issuing warning messages (non-critical alerts) to developers.
from typing import cast
# → Imports the cast() function from Python's typing module for explicit type casting in type hints.
# What it does:
# cast() tells type checkers (mypy, Pyright) to treat a value as a specific type, without any runtime effect.
from urllib.parse import urlparse
# → Imports the urlparse() function from Python's urllib.parse module for parsing URLs into components.
# What it does:
# urlparse() breaks a URL into its constituent parts for easy access and manipulation.
import dj_database_url
# → Imports the dj-database-url library, a popular Django utility for parsing database URLs from environment variables.
# What it does:
# Converts a database URL string into Django's database configuration dictionary.
import dj_email_url
# → Imports the dj-email-url library, a Django utility for configuring email settings from a single URL string.
# What it does:
# Parses an email URL and converts it into Django's email configuration dictionary
import django_cache_url
# → Imports the django-cache-url library, a Django utility for configuring cache backends from a single URL string.
# What it does:
# Parses a cache URL and converts it into Django's cache configuration dictionary.
import django_stubs_ext
# → Imports the django-stubs-ext extension package that enhances Django with better type hints for static type checkers (like mypy, Pyright).
# What it does:
# Monkey-patches Django to improve type inference and compatibility with Python type hints.
import sentry_sdk
# → Imports the Sentry SDK (Software Development Kit) for error tracking and performance monitoring.
# What is Sentry?
# Sentry is an error tracking service that helps developers monitor and fix crashes in real-time.
import sentry_sdk.utils
# → Imports utility modules from the Sentry SDK that provide helper functions and classes for error handling and data processing.
from celery.schedules import crontab
# → Imports the crontab class from Celery for defining periodic task schedules using cron-like syntax.
# What it does:
# Creates schedule objects that tell Celery when to run periodic tasks, similar to Linux cron jobs.
from django.conf import global_settings
# → Imports Django's global settings module, which contains the default values for all built-in Django settings.
# What it contains:
# global_settings is a module that defines the default values for every Django setting. These defaults are used unless overridden in your project's settings.py.
from django.core.cache import CacheKeyWarning
# → Imports the CacheKeyWarning exception class from Django's caching framework, which is raised when there are issues with cache keys.
# What it is:
# CacheKeyWarning is a warning (not an error) that Django issues when cache keys might cause problems.
from django.core.exceptions import ImproperlyConfigured
# → Imports Django's ImproperlyConfigured exception, which is raised when a Django application is configured incorrectly.
# What it is:
# ImproperlyConfigured is a core Django exception used to indicate configuration errors in Django applications.
from django.core.management.utils import get_random_secret_key
# → Imports a Django utility function that generates a cryptographically secure random secret key for Django projects.
# What it does:
# Generates a random 50-character string suitable for use as Django's SECRET_KEY setting.
from django.core.validators import URLValidator
# → Imports Django's URL validation class for checking if strings are valid URLs.
# What it does:
# URLValidator validates that a given string is a properly formatted URL according to RFC standards.
from graphql.execution import executor
# → Imports the executor module from GraphQL's core execution logic, which handles the runtime execution of GraphQL queries.

# What it does:
# The executor is the core engine that takes a GraphQL query and executes it against your schema, resolving fields and returning the result.
from pytimeparse import parse
# → Imports the parse function from the pytimeparse library, which converts human-readable time strings into seconds.
# What it does:
# Converts time expressions like "2 hours" or "3 days 4 hours" into the equivalent number of seconds.
from sentry_sdk.integrations.celery import CeleryIntegration
# → Imports the Celery integration for Sentry, which automatically captures errors and performance data from Celery tasks.
# What it does:
# Integrates Sentry error tracking with Celery distributed task queue, automatically capturing:
# Task failures and exceptions
# Task execution timing
# Retry attempts
# Task arguments and context
from sentry_sdk.integrations.django import DjangoIntegration
# → Imports the Django integration for Sentry, which automatically captures errors and performance data from Django applications.
# What it does:
# Integrates Sentry error tracking with Django, automatically capturing:
# Unhandled exceptions in views
# Database queries
# Template rendering errors
# Request/response data
# User information
# Performance metrics
from sentry_sdk.integrations.logging import ignore_logger
# → Imports a utility function that prevents specific loggers from sending events to Sentry.
# What it does:
# ignore_logger() tells Sentry to ignore all log messages from a specific logger, preventing them from being sent as Sentry events.
from sentry_sdk.scrubber import DEFAULT_DENYLIST, DEFAULT_PII_DENYLIST, EventScrubber
# → Imports Sentry's data scrubbing utilities for removing sensitive information from events before sending to Sentry.
# What they do:
# These utilities automatically redact sensitive data (passwords, tokens, credit card numbers, etc.) from Sentry events to prevent exposure of Personally Identifiable Information (PII).
# Components:
# 1. EventScrubber
# The main class that scrubs sensitive data from Sentry events.
# 2. DEFAULT_DENYLIST
# List of default sensitive fields to scrub (basic security-sensitive fields).
# 3. DEFAULT_PII_DENYLIST
# Extended list including PII fields (personal information).
from . import PatchedSubscriberExecutionContext, __version__
# → Imports two items from the current package/directory (relative import).
# What it does:
# . = Current package/directory (same folder as the importing module)
# PatchedSubscriberExecutionContext = A customized or modified version of a GraphQL subscriber execution context
# __version__ = Version string for the current package (e.g., "1.2.3")
from .account.i18n_rules_override import i18n_rules_override
# → Imports the i18n_rules_override object from the i18n_rules_override module within the account subpackage of the current package.
# What it does:
# This is a relative import that brings in internationalization (i18n) rule overrides for account-related functionality.
from .core.cleaners.html import HtmlCleanerSettings
# → Imports the HtmlCleanerSettings class from the html module within the cleaners subpackage of the core package in the current directory.
